add a bunch of confusing channel stuff
... in order to have the fsnotify stuff not interfere with a running
process e.g. git log or git blame

the important thing is we introduce a select statement in the main loop:

for {
    select {
    case <-watchChan:
        // display status update
    case <-inputChan:
        // previous body of the for loop would go here
    }
}

but in order to have a meaningful diff and not indent the entire file
one level, I instead use a goto (which could be a continue) to go to
the next iteration in the case of the fsnotify event

the other lines added are just to make it work right, but I acknowledge
that they are visual noise. this solution could be refined.